Update collision playground
