+++ tests/gdx-tests/src/com/badlogic/gdx/tests/utils/GdxTests.java
@@ -82,7 +82,6 @@ import com.badlogic.gdx.tests.gles2.SimpleVertexShader;
 import com.badlogic.gdx.tests.gles2.VertexArrayTest;
 import com.badlogic.gdx.tests.gles3.InstancedRenderingTest;
 import com.badlogic.gdx.tests.gles3.PixelBufferObjectTest;
-import com.badlogic.gdx.tests.math.collision.CollisionPlaygroundTest;
 import com.badlogic.gdx.tests.math.collision.OrientedBoundingBoxTest;
 import com.badlogic.gdx.tests.net.NetAPITest;
 import com.badlogic.gdx.tests.superkoalio.SuperKoalio;
@@ -130,7 +129,6 @@ public class GdxTests {
 		CollectionsTest.class,
 		CollisionPlaygroundTest.class,
 		ColorTest.class,
-		CollisionPlaygroundTest.class,
 		ContainerTest.class,
 		CoordinatesTest.class,
 		CpuSpriteBatchTest.class,
